fix(packages/viewer) export of viewer component
