+++ tsconfig.base.json
@@ -17,7 +17,7 @@
     "paths": {
       "@vctrl/hooks": ["packages/hooks/src/index.ts"],
       "@vctrl/hooks/*": ["packages/hooks/src/*"],
-      "@vctrl/viewer": ["packages/viewer/src/index.tsx"],
+      "@vctrl/viewer": ["packages/viewer/src/index.ts"],
       "@vctrl/viewer/*": ["packages/viewer/src/*"],
       "@vctrl/official-website/*": ["apps/official-website/src/*"],
       "@vctrl/shared/*": ["shared/src/*"]
